Switch to ValueState
